+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* `DiversityFeedScorer` also penalizes trending tags that show up more than `Config.minTrendingTagTootsForPenalty` times
 * Rename `Config.homeTimelineBatchSize`
 * Automatically serialize or deserialize toots to/from `Storage.get/set` based on values in `STORAGE_KEYS_WITH_TOOTS`
 * Configure STORAGE_KEYS_WITH_ACCOUNTS, STORAGE_KEYS_WITH_UNIQUE_IDS, call unique checker in debug mode loading from Storage